refactor(ai-builder): Modularize pairwise evaluation using LangSmith best practices
Split pairwise-runner.ts (~980 lines) into focused modules:

- pairwise-generator.ts: Target function for LangSmith evaluate()
  Uses traceable wrapper with getLangchainCallbacks() for proper tracing

- pairwise-ls-evaluator.ts: LangSmith evaluator that runs judge panel
  Separates generation from evaluation per LangSmith patterns

- utils/judge-panel.ts: Shared judge panel execution and aggregation
  Exports runJudgePanel(), aggregateJudgeResults(), aggregateGenerations()

- utils/artifact-saver.ts: Extracted artifact persistence logic

Key improvements:
- Use LangSmith's numRepetitions instead of manual createRepeatedData()
- Remove global mutable state (generationCounter)
- Clear separation: generator outputs workflow, evaluator scores it
- Main file reduced from ~980 to ~380 lines

Note: Multi-generation mode currently evaluates single workflow from
generator. TODO to restore full multi-gen by regenerating in evaluator.
